# Conflicts:
#	Server/xingzhen-xz/xingzhen-xz-api/src/main/java/com/hisign/xingzhen/xz/api/entity/Ajgroup.java

+++ Server/xingzhen-xz/xingzhen-xz-service/src/main/java/com/hisign/xingzhen/xz/rest/TaskRest.java
@@ -2,6 +2,8 @@ package com.hisign.xingzhen.xz.rest;
 
 import com.hisign.bfun.bif.BaseRest;
 import com.hisign.bfun.bmodel.JsonResult;
+import com.hisign.bfun.butils.JsonResultUtil;
+import com.hisign.xingzhen.common.util.StringUtils;
 import com.hisign.xingzhen.xz.api.entity.Task;
 import com.hisign.xingzhen.xz.api.model.TaskModel;
 import com.hisign.xingzhen.xz.api.param.TaskAddParam;
@@ -40,7 +42,7 @@ public class TaskRest extends BaseRest<Task, TaskModel, String, TaskService> imp
     @Override
     @ApiOperation(value = "任务管理查询分页",httpMethod ="POST",response = TaskModel.class)
     @RequestMapping(value = "/getTaskPage", method = RequestMethod.POST, produces = {"application/json;charset=UTF-8"})
-    public JsonResult getTaskPage(@ApiParam @RequestBody TaskSelectParam taskSelectParam) {
+    public JsonResult getTaskPage(@RequestBody TaskSelectParam taskSelectParam) {
         return baseService.getTaskPage(taskSelectParam);
     }
 
@@ -50,9 +52,12 @@ public class TaskRest extends BaseRest<Task, TaskModel, String, TaskService> imp
      * @return
      */
     @Override
-    @ApiOperation(value = "任务详情",httpMethod ="GET",response = String.class)
-    @RequestMapping(value = "/taskDetail", method = RequestMethod.GET, produces = {"application/json;charset=UTF-8"})
+    @ApiOperation(value = "任务详情",httpMethod ="POST",response = TaskModel.class)
+    @RequestMapping(value = "/taskDetail", method = RequestMethod.POST, produces = {"application/json;charset=UTF-8"})
     public JsonResult taskDetail(@RequestParam String id,@RequestParam String userId) {
+        if(StringUtils.isEmpty(userId)){
+            return JsonResultUtil.error("查看记录失败,当前登陆用户不能为空");
+        }
         return baseService.taskDetail(id, userId);
     }
 
@@ -64,7 +69,10 @@ public class TaskRest extends BaseRest<Task, TaskModel, String, TaskService> imp
     @Override
     @ApiOperation(value = "新增任务",httpMethod ="POST",response = String.class)
     @RequestMapping(value = "/addTask", method = RequestMethod.POST, produces = {"application/json;charset=UTF-8"})
-    public JsonResult addTask(@ApiParam @RequestBody TaskAddParam taskAddParam) {
+    public JsonResult addTask(@RequestBody TaskAddParam taskAddParam) {
+        if(StringUtils.isEmpty(taskAddParam.getCreator()) || StringUtils.isEmpty(taskAddParam.getCreatename())){
+            return JsonResultUtil.error("添加记录失败,当前登陆用户不能为空");
+        }
         return baseService.addTask(taskAddParam);
     }
 
@@ -74,10 +82,12 @@ public class TaskRest extends BaseRest<Task, TaskModel, String, TaskService> imp
      * @return
      */
     @Override
-    @ApiOperation(value = "删除任务",httpMethod ="DELETE",response = String.class)
-    @ApiImplicitParams({ @ApiImplicitParam(name="id",value = "任务id",required = true,dataType = "String"),@ApiImplicitParam(name="userId",value = "当前用户id",required = true,dataType = "String")})
-    @RequestMapping(value = "/deleteTaskById", method = RequestMethod.DELETE, produces = {"application/json;charset=UTF-8"})
+    @ApiOperation(value = "删除任务",httpMethod ="POST",response = String.class)
+    @RequestMapping(value = "/deleteTaskById", method = RequestMethod.POST, produces = {"application/json;charset=UTF-8"})
     public JsonResult deleteTaskById(@RequestParam String id,@RequestParam String userId) {
+        if(StringUtils.isEmpty(userId)){
+            return JsonResultUtil.error("删除记录失败,当前登陆用户不能为空");
+        }
         return baseService.deleteTaskById(id,userId);
     }
 
@@ -89,7 +99,10 @@ public class TaskRest extends BaseRest<Task, TaskModel, String, TaskService> imp
     @Override
     @ApiOperation(value = "移交任务",httpMethod ="POST",response = String.class)
     @RequestMapping(value = "/moveTask", method = RequestMethod.POST, produces = {"application/json;charset=UTF-8"})
-    public JsonResult moveTask(@ApiParam @RequestBody TaskMoveParam taskMoveParam) {
+    public JsonResult moveTask(@RequestBody TaskMoveParam taskMoveParam) {
+        if(StringUtils.isEmpty(taskMoveParam.getCreator()) || StringUtils.isEmpty(taskMoveParam.getCreatename())){
+            return JsonResultUtil.error("任务移交失败,当前登陆用户不能为空");
+        }
         return baseService.moveTask(taskMoveParam);
     }
 }
